moving link functions to separate file
