+++ handlers.go
@@ -65,21 +65,7 @@ func (s *HTTPServer) ApiGetFileSections(w http.ResponseWriter, r *http.Request) 
 	writeJsonResponse(fileContents, w, r)
 }
 
-type LinksFromFile struct {
-	Title string         `json:"title"`
-	Links []MarkdownLink `json:"links"`
-}
-
 func (s *HTTPServer) ApiGetAllLinks(w http.ResponseWriter, r *http.Request) {
-
-	var allLinks []LinksFromFile
-	for _, f := range s.Index.Files {
-		lff := LinksFromFile{
-			Title: f.Title,
-			Links: f.GetLinks(),
-		}
-		allLinks = append(allLinks, lff)
-	}
-
+	allLinks := s.Index.GetAllLinks()
 	writeJsonResponse(allLinks, w, r)
 }
